Eye Blink Detection/requirements.txt
streamlit==1.32.0
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
imutils==0.5.4 